requirements.txt
asgiref==3.4.1
autopep8==1.6.0
boto3==1.19.8
botocore==1.22.8
defusedxml==0.7.1
diff-match-patch==20230430
Django==3.2.9
django-cors-headers==3.10.0
django-filter==21.1
django-import-export==3.2.0
django-model-utils==4.2.0
django-storages==1.12.3
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
gunicorn==20.1.0
inflection==0.5.1
jmespath==0.10.0
Markdown==3.3.4
MarkupPy==1.14
mssql-django==1.3
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
pycodestyle==2.8.0
PyJWT==2.8.0
PyMySQL==1.1.0
pyodbc==5.0.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
tablib==3.5.0
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.7
xlrd==2.0.1
xlwt==1.3.0